:recycle: configを整理
