+++ src/config.rs
@@ -1,2 +1,6 @@
 pub mod settings;
 pub mod database_settings;
+pub mod embedding_model;
+pub mod vector_memory_config;
+pub mod config_loader;
+pub mod env_loader;
