+++ src/repositories/vector_memory_repository.rs
@@ -1,46 +1,9 @@
 use uuid::Uuid;
 use async_trait::async_trait;
+use crate::config::vector_memory_config::VectorMemoryConfig;
 use crate::errors::custom::CustomError;
 use crate::models::internal::memory_entry::MemoryEntry;
 use crate::models::public::memory_filters::MemoryFilters;
-
-/// Available embedding models and their corresponding vector sizes
-#[derive(Debug, Clone, Copy)]
-pub enum EmbeddingModel {
-    /// OpenAI text-embedding-3-small model (1536 dimensions)
-    TextEmbedding3Small,
-    /// OpenAI text-embedding-3-large model (3072 dimensions)
-    TextEmbedding3Large,
-    /// OpenAI text-embedding-ada-002 model (1536 dimensions)
-    TextEmbeddingAda002,
-}
-
-/// Configuration for VectorMemoryRepository
-///
-/// * `url` - URL of the vector database server
-/// * `collection_name` - Name of the collection to store memories
-/// * `vector_size` - Size of the embedding vectors
-#[derive(Debug, Clone)]
-pub struct VectorMemoryConfig {
-    pub(crate) url: String,
-    pub(crate) collection_name: String,
-    pub(crate) vector_size: u64,
-}
-
-impl VectorMemoryConfig {
-    pub(crate) fn new(url: String, collection_name: String, model: EmbeddingModel) -> Self {
-        let vector_size = match model {
-            EmbeddingModel::TextEmbedding3Small => 1536,
-            EmbeddingModel::TextEmbedding3Large => 3072,
-            EmbeddingModel::TextEmbeddingAda002 => 1536,
-        };
-        Self {
-            url,
-            collection_name,
-            vector_size,
-        }
-    }
-}
 
 /// Repository trait for storing and retrieving memories using a vector database.
 #[cfg_attr(test, mockall::automock)]
